refactor : 접근 제어자 선언 및 메소드 위치 변경
